HW TestNG

+++ src/main/java/Calculator.java
@@ -1,14 +1,18 @@
 public class Calculator {
-public int sum (int a, int b) {
-    return a+b;
-}
-public int difference (int a, int b) {
-    return a-b;
-}
-public int multiplication (int a, int b) {
-    return a*b;
-}
-public int division (int a, int b) {
-    return a/b;
-}
+
+    public int sum(int a, int b) {
+        return a + b;
+    }
+
+    public int difference(int a, int b) {
+        return a - b;
+    }
+
+    public int multiplication(int a, int b) {
+        return a * b;
+    }
+
+    public double division(double a, double b) {
+        return a / b;
+    }
 }
